CSC450-Circuit/requirements.txt
sqlite3
flask
werkzeug
python-dotenv